requirements.txt
fastapi
uvicorn
python-dotenv
torch
clip-by-openai
pillow
requests
faiss-cpu
transformers
timm